Updated tests for pull and update_dict
